requirements.txt
fastapi==0.70.0
gunicorn==20.1.0
image==1.5.33
imageio==2.22.2
imgaug==0.4.0
numpy==1.22.4
opencv-python==4.5.4.60
packaging==21.3
paddle-bfloat==0.1.7
paddle2onnx==0.8.1
paddlefsl==1.1.0
paddlehub==2.3.0
paddlenlp==2.4.1
paddlepaddle==2.3.2
paddlepaddle-gpu==2.3.2
pandas==1.5.0
parso==0.8.3
Pillow==8.4.0
SQLAlchemy==1.4.27
uvicorn==0.15.0
